Change of settings is done in main.go

+++ generator.go
@@ -5,12 +5,12 @@ import (
 	"time"
 )
 
-func generator(output chan<- datablock, max int, rate float64) {
+func generator(output chan<- datablock, max int, rate int) {
 	ts := rand.NewSource(time.Now().UnixNano())
 	r := rand.New(ts)
 	d := datablock{}
 
-	ticker := time.NewTicker(time.Duration(1/rate*1000000000) * time.Nanosecond)
+	ticker := time.NewTicker(time.Duration(1/float64(rate)*1000000000) * time.Nanosecond)
 	for {
 		select {
 		case d.timestamp = <-ticker.C:
